Implementation of PermSweep
Initial Implementation of PermSweep - Permutation Sweep.  When the GA reaches the generation-stall halt conditions, the fittest solution is passed to the PermSweep algorithm, which analyzes every single-step solutions derived from it; it generates all possible solutions that differ from the current best by a single (flipped) bit.  These new solutions' fitnesses are compared with the orignal solution, and if any of the new solutions improve on the original, it becomes the new best fit, resets the halting counts, and restarts the GA.

Currently only sweeps, it does not perform the permutative part of that sweep.  The idea is to randomize the order of the sweep as the fitness sorting is inherently stable, and will favor BSs early in the standard BS order; if the order is randomized, then there is no consistent favorability.