PermSweep Injection Correction
Correction to PermSweep algorithm to inject any potential new member into the member pool by replacing the least fit member, not kicking out the last member of the generation (which was incorrectly assumed to BE the least fit member of that generation; the member list isn't sorted until the START of the next generation).

Also synchronizes temp.dat and *_Setup.m between machines.