Plot Maximal Allocation Improvement
Because of rounding, BSs that are fully allocated are "over" allocated by less than 0.1% (perhaps even lower), and thusly are plotted as if they are overallocated.  Adds a small fix to the plotting function that adjusts the "over" allocated figures to be BSs allocated at >100.1% maximum allocation (rather than at >100%).