Alpha Fix
Easiest possible fix: apply the normalization factor (the sum rate of the field) to alpha within MATLAB and report the factor as 1 to CPLEX.  May be insufficient, and might need to be applied to del_mso, instead; that is next attempt.

Applies a change to the writing functions from 'w' mode to 'W'.  This tells MATLAB to not empty the buffer, which should decrease the write time of the CPLEX output files without causing any change to the written files.

Applies a change to the halting condition of the GA, so that the member halt (halt when fittest member hasn't changed in 'y' generations) only counts up while the overage cost is in excess of the most expensive member.