Parse/Plot Update; Data Formatting; Sync
Update to data parsing to properly pull from the files, determine satisfaction according to the results, store the data within a struct, and save the struct (and associated range) within an appropriate MATLAB data file (*.mat) according to the range.

Corrects VorMod.m to recombine data writing to the end of the genetic algorithm.

Data Synchronization