File and Data Formatting Updates
Began transition of VAE to VAS (evaluating the "solution", or training set is now VAS while VAE is reserved for new data sets for pairity with VOE).

Removed references for cpfn and derivatives.  Everything uses the local folder as root, so cpfn only included a single folder, and scripts were inconsistent regarding use.

Added parsing support for VAS, and began support for new demand point generation to be soon used with VAE and VOE.