GA.onegeneration Function Complete; Fixes
Completed GA.onegeneration() and the accompanying functions for processing a single generation.  Moves through selection, crossover, mutation, and the uniqueness check to derive the next generation.  Further, updated VMGA.computefitness to appropriately derive fitness value for nGenerations ~= 0.

Note: while GA.onegeneration works, it's performace is oddly abysmal compared to the original implementation.  Something about VMGA.computefitness takes a horrendous amount of time to process compared to the orginal implementation, despite being functionally identical.  Profiling idicates that it takes the vast majority of the time (4.17 out of 4.32 seconds for the whole GA.onegeneration method which calls it) to process.  Work needs to continue on this function to optimize, or else this entire refactoring effort may be a waste.

Additional note: Accidentally fucked up my commit earlier, causing loss of implementation.  Was able to recover as a result of still having the initial matlab .asv autosave file open, so was able to recover most of the commit (GA only) and re-implemented the rest (VMGA and one line in VorMod.m).  Saving the bak_* file I created out of the matlab .asv file. Just in case.