Corrected Fitness Calculation; Applied to All-On Test
Original finess calculation was applying the base station load to the incorrect base station; this was due to an array mismatch where it would save to base stations that were listed within the baseStationIndices vector, while looking up their load within the obj.demandField matrix, which specified the base station a given pixel was closest to via the base station's index in baseStationIndicies.  Effectively the matlab equivalent to a pointer error.  Corrected, and now the new, fast method is accurate to the old method within 1e-9 (or approximately 14 or 15 orders of magnitude, compared to the values being at 1e5 or 1e6).

Applied the new method to deriving the all-on, beta = 1 scenario (in VorMod.m) and removed the old method's comments within the computefitness method of VorModGeneticAlgorithm.m.