Public Beta; findfittestmembers varargout
Beta now has public set access, as it will need to be modified for subsequent runs.  Will probably need a resetmembers method, as well.

Added comments to VMGA as well to detail the VMGA-specific properties.

Modified GA findfittestmembers method to have a varargout that also outputs the fittest members index in addition to the members themselves.  Might want to also output the member fitness.