Fixes, Reporting, Compute Solution
Recovered from crash; there was virtually no damage that could have been recovered using the .asv's - they were older than the saved files.

Updated GA.m to use int32s instead of int8s (was causing overflows), and not pluralize min and maxGeneration.

Updated VMGA to include documentation, changed names of the flags from flagReport and flagReportGraph to flagGenerationReport and flagGenerationReportGraph, and implemented the shouldhalt method with associated properties and implications into the showgenerationreport method.  Also updated VMGA to store the cost and overage of the fittest member, to expedite halt and showgenerationreport.

Updated VorMod.m for legibility, appropriate calling of VMGA, and controlling and constants.