Added Reset Functions
Added reset functionality to GA and VMGA in order purge the GA and start over with (perhaps) different settings.

Also a setter function for the beta parameter so it updated the demand field parameter accordingingly.  This caused a warning to be flagged regarding to how additional parameters were referenced within the function; this may cause problems and should be looked into.

Need to profile run time.  Seemed to run slower now that the full-processing loop was created.