Data Writing
Expanded VorMod to compute all solutions as indicated in Beta, save the data in a new object (GeneticAlgorithmSolution), and write that data to file.

GeneticAlgorithmSolution is a basic object to hold all the important data output by the GeneticAlgorithm class.  It's generic, and should be sufficient for any GeneticAlgorithm child, and could be expanded.

Wrote generic functions for writing data files for two-stage (cplexWriteDataTwoStage.m) and one-stage (cplexWriteDataOneStage.m) voronoi cplex models.

Corrected VorModGeneticAlgorithm class to ignore possible issues with the beta setter.  Also added functionality for data tracking and fixed small problems with resetting with the VMGA and GeneticAlgorithm.