File Writer Buffer
Writing one of the .dat files had missing data.  Files are smaller, so changed writer from 'W' mode to 'w' mode.  This clears the buffer, and should ensure that doesn't happen. (I think; worth a try).

Original change was to expedite file writing.  'W' is faster than 'w', so was changed, but this doesn't flush the buffer.