Implemented Solution-Stage Data Reading
Implemented one- and two-stage vormod cplex solution reading as separate objects.  Both parse the original .dat file that initializes the cplex model so that it does not need the matlab environment to run, and parses the assorted created cplex output files for contained data.  Also created a function to automate creation of these objects and saves them to a matlab archive for expedited reading.

Made assorted corrections:
cplexWriteDataOneStage to use the correct number of dimensions for rateNorm and formatting for activeResources.
Vormod to actually call cplexWriteDataOneStage correctly - that is by providing the active base stations and not just the beta value of the solution meant to be solved.