Added VMGA Plotting
Add method for plotting a given member in VMGA.  Uses the old gradient plot method, applied directly within the class rather than as an outside function; just need to pass the member to plot.  The control variable for graphing has been updated from a strict boolean to a nonnegative integer which is the handle of the figure to draw to.

Modified VorMod.m to synchronize data, increased the maximum number of generations of the genetic algorith, correct the graphing control variable to be of the correct form, and correct writing rateNormalization in VOS and VAS data sets.

Fixed error in DemandRealizations class which has been left during previous modifications.